rewrited pong in vue component
